Support --socket-timeout
